requirements/dev.txt
-r base.txt

model-bakery
django-debug-toolbar==4.3.0
faker